# Backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "climate-cigarette-tracker-api",
    "description": "This Backend will Handle all the user request to get Amount of cigarette consume due to pollution and recommend prevenstion or todo etc.",
    "version": "1.0.0"
  },
  "host": "shy-crab-swimsuit.cyclic.app",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/logs/downloads/application.log": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/logs/downloads/error.log": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/send/location": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "lat": {
                  "example": "any"
                },
                "lon": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/calculate": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "aqi": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}